Change a typ in the MOD interface,changed the calculation of the Speedmod ->now supports stupid and inteligend mode,RecipeMod options and Custom Recipes

+++ java/me/zero/cc/Zero_lite/Mods/ChatMod.java
@@ -95,7 +95,7 @@ public class ChatMod implements Mod {
 	}
 
 	@Override
-	public void manupulateValue(String ValueToManupulate, int percent) {
+	public void manupulateValue(String ValueToManupulate, double percent) {
 		
 	}
